Move sub-classes of Transport to transport.py.
